@@ -19,9 +19,9 @@ export class FirebaseAppConfig {
   apiKey?: string;
   authDomain?: string;
   databaseURL?: string;
+  projectId?: string;
   storageBucket?: string;
   messagingSenderId?: string;
-  projectId?: string;
 }
 
 @NgModule({
